Rename: info, review -> infoSection, reviewSection (#1)
