+++ src/components/detail/InfoSection/InfoSection.module.css
